+++ src/main/java/com/aliens/backend/email/service/EmailService.java
@@ -3,7 +3,7 @@ package com.aliens.backend.email.service;
 import com.aliens.backend.email.controller.response.EmailResponse;
 import com.aliens.backend.email.domain.EmailAuthentication;
 import com.aliens.backend.email.domain.repository.EmailAuthenticationRepository;
-import com.aliens.backend.global.encode.SymmetricKeyEncoder;
+import com.aliens.backend.member.sevice.SymmetricKeyEncoder;
 import com.aliens.backend.global.error.EmailError;
 import com.aliens.backend.global.exception.RestApiException;
 import com.aliens.backend.member.controller.dto.event.TemporaryPasswordEvent;
