+++ src/main/java/com/aliens/backend/member/sevice/MemberInfoService.java
@@ -6,7 +6,6 @@ import com.aliens.backend.auth.domain.repository.MemberRepository;
 import com.aliens.backend.auth.service.PasswordEncoder;
 import com.aliens.backend.email.domain.EmailAuthentication;
 import com.aliens.backend.email.domain.repository.EmailAuthenticationRepository;
-import com.aliens.backend.global.encode.SymmetricKeyEncoder;
 import com.aliens.backend.global.error.EmailError;
 import com.aliens.backend.global.error.MemberError;
 import com.aliens.backend.global.exception.RestApiException;
@@ -20,7 +19,7 @@ import com.aliens.backend.member.controller.dto.request.SignUpRequest;
 import com.aliens.backend.member.domain.*;
 import com.aliens.backend.member.domain.repository.MemberInfoRepository;
 import com.aliens.backend.uploader.AwsS3Uploader;
-import com.aliens.backend.uploader.S3File;
+import com.aliens.backend.uploader.dto.S3File;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
