+++ src/main/java/com/aliens/backend/uploader/dto/S3File.java
@@ -1,4 +1,4 @@
-package com.aliens.backend.uploader;
+package com.aliens.backend.uploader.dto;
 
 public record S3File(String fileName, String fileURL) {
 }
