+++ src/test/java/com/aliens/backend/member/MemberInfoServiceTest.java
@@ -6,7 +6,7 @@ import com.aliens.backend.auth.domain.repository.MemberRepository;
 import com.aliens.backend.auth.service.PasswordEncoder;
 import com.aliens.backend.email.domain.EmailAuthentication;
 import com.aliens.backend.email.domain.repository.EmailAuthenticationRepository;
-import com.aliens.backend.global.encode.SymmetricKeyEncoder;
+import com.aliens.backend.member.sevice.SymmetricKeyEncoder;
 import com.aliens.backend.member.controller.dto.*;
 import com.aliens.backend.member.controller.dto.request.SignUpRequest;
 import com.aliens.backend.member.controller.dto.request.TemporaryPasswordRequest;
@@ -19,7 +19,7 @@ import com.aliens.backend.member.domain.repository.MemberInfoRepository;
 import com.aliens.backend.member.domain.MemberStatus;
 import com.aliens.backend.member.sevice.MemberInfoService;
 import com.aliens.backend.uploader.AwsS3Uploader;
-import com.aliens.backend.uploader.S3File;
+import com.aliens.backend.uploader.dto.S3File;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
